Economic/immoscoop.be/requirements/requirements.txt
tesseract
pytesseract
Image
Pillow
requests
lxml
elasticsearch
kafka-python
